replace whole value instead of only matches for relabeling
Change the behavior for the dynamic relabeling so that for a match the
whole value is replaced by the replacement instead of the matches within
the value being replaced by the replacement.

This way it is possible to use regular expressions for matching that
don't cover the whole value but still replace it completely.